Improve the random number generator
improved the random number generator to use time(NULL) instead of clock(), as the seed. This makes it less deterministic, as I am now using the systems time (in seconds) to get a random number, instead of the programs clock.

Additionally, I changed the bool data types to use a enum instead of macros, as this allows them to be used much more like other high level languages.